refactor: extract BreakdownTable, ConfidenceSlider, and SearchResultItem
Closes #2058
Closes #2032
Closes #2060
Closes #2033
Closes #2044
Closes #2017

- Extract inline BreakdownTable from AiUsagePage to app-ai/components/BreakdownTable.tsx
- Extract inline ConfidenceSlider from RulesBrowserPage to app-ai/components/ConfidenceSlider.tsx
- Add SearchResultItem composite to @pops/ui with leading/title/meta/trailing slots
- Refactor all 6 domain search result components (finance, inventory, media) to use SearchResultItem
