+++ packages/app-ai/src/pages/RulesBrowserPage.tsx
@@ -1,5 +1,5 @@
 import { BookOpen, Trash2 } from 'lucide-react';
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 import { trpc } from '@pops/api-client';
 /**
@@ -24,11 +24,11 @@ import {
   Select,
   type SelectOption,
   Skeleton,
-  Slider,
   SortableHeader,
   TextInput,
-  useDebouncedCallback,
 } from '@pops/ui';
+
+import { ConfidenceSlider } from '../components/ConfidenceSlider';
 
 import type { ColumnDef } from '@tanstack/react-table';
 
@@ -57,73 +57,6 @@ const MATCH_TYPE_OPTIONS: SelectOption[] = [
 ];
 
 const PAGE_SIZE = 50;
-
-function ConfidenceSlider({
-  ruleId,
-  initial,
-  onAutoDelete,
-}: {
-  ruleId: string;
-  initial: number;
-  onAutoDelete: (id: string) => void;
-}) {
-  const [value, setValue] = useState(initial);
-  const initialRef = useRef(initial);
-  const utils = trpc.useUtils();
-
-  // Keep initialRef in sync when the prop changes (e.g. after query invalidation)
-  useEffect(() => {
-    initialRef.current = initial;
-    setValue(initial);
-  }, [initial]);
-
-  const adjustMutation = trpc.core.corrections.adjustConfidence.useMutation({
-    onSuccess: () => {
-      void utils.core.corrections.list.invalidate();
-    },
-  });
-
-  const commit = useCallback(
-    (newValue: number) => {
-      const delta = newValue - initialRef.current;
-      if (Math.abs(delta) < 0.001) return;
-      adjustMutation.mutate(
-        { id: ruleId, delta },
-        {
-          onSuccess: () => {
-            if (newValue < 0.3) {
-              onAutoDelete(ruleId);
-            }
-          },
-        }
-      );
-    },
-    [ruleId, adjustMutation, onAutoDelete]
-  );
-
-  const debouncedCommit = useDebouncedCallback(commit, 400);
-
-  const handleChange = (values: number[]) => {
-    const next = values[0] ?? value;
-    setValue(next);
-    debouncedCommit(next);
-  };
-
-  return (
-    <div className="flex items-center gap-2 min-w-35">
-      <Slider
-        min={0}
-        max={1}
-        step={0.01}
-        value={[value]}
-        onValueChange={handleChange}
-        className="w-20"
-        aria-label={`Confidence for rule ${ruleId}`}
-      />
-      <span className="text-xs tabular-nums w-10 text-right">{(value * 100).toFixed(0)}%</span>
-    </div>
-  );
-}
 
 export function RulesBrowserPage(): React.ReactElement {
   const [matchType, setMatchType] = useState('');
