+++ packages/app-finance/src/components/search/TransactionsResultComponent.tsx
@@ -1,5 +1,5 @@
 import { registerResultComponent } from '@pops/navigation';
-import { formatCurrency, formatDate, highlightMatch } from '@pops/ui';
+import { formatCurrency, formatDate, highlightMatch, SearchResultItem } from '@pops/ui';
 
 import type { ResultComponentProps } from '@pops/navigation';
 
@@ -27,26 +27,22 @@ export function TransactionsResultComponent({ data, query, matchField }: ResultC
   const shouldHighlight = matchField === 'description' && query;
 
   return (
-    <div className="flex items-center justify-between gap-2 min-w-0">
-      <div className="flex flex-col min-w-0">
-        <span className="text-sm font-medium truncate">
-          {shouldHighlight ? highlightMatch(tx.description, query) : tx.description}
-        </span>
-        {tx.entityName && (
-          <span className="text-xs text-muted-foreground truncate">{tx.entityName}</span>
-        )}
-      </div>
-      <div className="flex flex-col items-end shrink-0">
-        <span className={`text-sm font-medium ${amountColorClass(tx.type)}`}>
-          {tx.type === 'income' ? '+' : tx.type === 'expense' ? '-' : ''}
-          {formatCurrency(Math.abs(tx.amount), {
-            minimumFractionDigits: 2,
-            maximumFractionDigits: 2,
-          })}
-        </span>
-        <span className="text-xs text-muted-foreground">{formatDate(tx.date)}</span>
-      </div>
-    </div>
+    <SearchResultItem
+      title={shouldHighlight ? highlightMatch(tx.description, query) : tx.description}
+      meta={tx.entityName ? [<span key="entity">{tx.entityName}</span>] : undefined}
+      trailing={
+        <div className="flex flex-col items-end shrink-0">
+          <span className={`text-sm font-medium ${amountColorClass(tx.type)}`}>
+            {tx.type === 'income' ? '+' : tx.type === 'expense' ? '-' : ''}
+            {formatCurrency(Math.abs(tx.amount), {
+              minimumFractionDigits: 2,
+              maximumFractionDigits: 2,
+            })}
+          </span>
+          <span className="text-xs text-muted-foreground">{formatDate(tx.date)}</span>
+        </div>
+      }
+    />
   );
 }
